:recycle: refactor: 커스텀 예외 삭제, 표준 예외 사용
