+++ src/main/java/lotto/validator/Validator.java
@@ -8,7 +8,7 @@ import static lotto.exception.ErrorMessage.WRONG_SIZE;
 
 import java.util.List;
 import lotto.domain.Lotto;
-import lotto.exception.LottoException;
+import lotto.exception.ErrorMessage;
 
 public class Validator {
     private static final int LOTTO_SIZE = 6;
@@ -19,32 +19,32 @@ public class Validator {
 
     public static void validateSize(List<Integer> numbers) {
         if (numbers.size() != LOTTO_SIZE) {
-            throw LottoException.of(WRONG_SIZE);
+            throw ErrorMessage.newIllegalArgumentException(WRONG_SIZE);
         }
     }
 
     public static void validateRange(List<Integer> numbers) {
         if (numbers.stream().anyMatch(Validator::measureRange)) {
-            throw LottoException.of(OUTSIDE_RANGE);
+            throw ErrorMessage.newIllegalArgumentException(OUTSIDE_RANGE);
         }
     }
 
     public static void validateDuplicate(List<Integer> numbers) {
         if (numbers.stream().distinct().count() != LOTTO_SIZE) {
-            throw LottoException.of(DUPLICATE_WINNING_NUMBERS);
+            throw ErrorMessage.newIllegalArgumentException(DUPLICATE_WINNING_NUMBERS);
         }
     }
 
     public static void validateUnique(int number, Lotto numbers) {
         if (numbers.getNumbers().contains(number)) {
-            throw LottoException.of(DUPLICATE_BONUS_NUMBER);
+            throw ErrorMessage.newIllegalArgumentException(DUPLICATE_BONUS_NUMBER);
         }
     }
 
     // LOTTO_PRISE를 한 곳에서 관리하기 위해 리턴 값 설정함
     public static int validateUnit(int purchaseAmount) {
         if (calculateRemainder(purchaseAmount) != PROPER_REMAINDER) {
-            throw LottoException.of(UNFIT_UNIT);
+            throw ErrorMessage.newIllegalArgumentException(UNFIT_UNIT);
         }
         return calculateUnit(purchaseAmount);
     }
